==> 4. Instacart Consistency Chekcs.ipynb <==
# import libraries
import pandas as pd
import numpy as np
import os

# import orders_csv
df_ords = pd.read_csv(r'/Users/jlsanabria77/Desktop/17-02.2021 Instacat Basket Analysis 3/02 Data/Original Data/orders.csv', index_col = False)

# import products_csv
df_prods = pd.read_csv(r'/Users/jlsanabria77/Desktop/17-02.2021 Instacat Basket Analysis 3/02 Data/Original Data/products.csv', index_col = False)

# import departments_csv
df_dep = pd.read_csv(r'/Users/jlsanabria77/Desktop/17-02.2021 Instacat Basket Analysis 3/02 Data/Original Data/departments.csv', index_col = False)

path = r'/Users/jlsanabria77/Desktop/17-02.2021 Instacat Basket Analysis 3'

path

df_dep.head()

# transposing df_dep
df_dep.T

# create new dataframe from the transposed version
df_dep_t = df_dep.T

df_dep_t

df_dep_t.reset_index()

# take the first row of df_dep_t for the header
new_header = df_dep_t.iloc[0]

new_header

# take the data under the header row for a new df
df_dep_t_new = df_dep_t[1:]

df_dep_t_new

# set the header row as the df header
df_dep_t_new.columns = new_header

df_dep_t_new

# turn dataframe into data dictionary
data_dict = df_dep_t_new.to_dict('index')

data_dict

df_prods.head()

print(data_dict.get('19'))

df_snacks =  df_prods[df_prods['department_id']==19]

df_prods['department_id']==19

df_snacks =  df_prods[df_prods['department_id']==19]

df_snacks.head()

# Beginning of Task for exercise 4.4

df_ords.head()

# 2. change a variable which doesnt need to be identified as numeric variable into a suitable format
df_ords['order_id'] = df_ords['order_id'].astype('str')

# check order_id new data type (object)
df_ords['order_id'].dtype

# 3. change unintuitive name for variable order_dow
df_ords.rename(columns = {'order_dow' : 'orders_day_of_the_week'}, inplace = True)

# check name has been changed
df_ords.head()

# 4. busiest hour for placing orders. DON'T KNOW HOW TO DO THAT
df_ords['order_hour_of_day'].value_counts(dropna = False)

The busiest hour for placing orders is 10am.

# 5. determine meaning of 4 in 'department_id' columm. We can use steps 31 and 32 from the exercise
data_dict = df_dep_t_new.to_dict('index')

# Meaning of department-id is "produce"
data_dict

# 6. subset for "breakfast" items
df_breakfast=df_prods[df_prods['department_id']==14]

df_breakfast

# 7. Details of customers & dinner parties (5-alcohol, 20-deli, 7-beverages, 12-meat seafood)
df_party=df_prods.loc[df_prods['department_id'].isin([5,20,7,12])]

df_party

8. last data frame created row count: 7650

# 9. Extract all information about "user_id" of 1
df_customer_1.describe()

Based on this dataframe we can get to the conclusion that this customer orders from Monday to Thursday and never on Fri, Sat or Sun. Aslo we can see "customer #1' has never ordered later than 4pm (16) most orders are made around 10am. Finally we can see those orders are very spaced out in time. The shortest number of days between orders ia 14 and the longest 30. 

df_ords.to_csv(os.path.join(path, '02 Data','Prepared Data', 'orders_wrangled.csv'))

df_dep_t_new.to_csv(os.path.join(path, '02 Data','Prepared Data', 'departments_wrangled.csv'))
