==> 5. Instacart Data Consistency Checks.ipynb <==
# Import libraries
import pandas as pd
import numpy as np
import os

# Turning folderpath into a string
path = r'/Users/jlsanabria77/Desktop/17-02.2021 Instacat Basket Analysis 3'

path

# Import "products_csv"
df_prods = pd.read_csv(os.path.join(path, '02 Data','Original Data','products.csv'), index_col = False)

# Import "orders_csv"
df_ords = pd.read_csv(os.path.join(path, '02 Data','Original Data','orders.csv'), index_col = False)

df_ords.describe()

# Create a dataframe
df_test = pd.DataFrame()

# Create a mixed type column
df_test['mix'] = ['a', 'b', 1, True]

df_test.head()

# Check for mixed types

for col in df_test.columns.tolist():
  weird = (df_test[[col]].applymap(type) != df_test[[col]].iloc[0].apply(type)).any(axis = 1)
  if len (df_test[weird]) > 0:
    print (col)

df_test['mix'] = df_test['mix'].astype('str')

df_prods.isnull().sum()

df_nan=df_prods[df_prods['product_name'].isnull()== True]

df_nan

df_prods.shape

df_prods_clean=df_prods[df_prods['product_name'].isnull()== False]

df_prods_clean.shape

df_dups = df_prods_clean[df_prods_clean.duplicated()]

df_dups

df_prods_clean.shape

df_prods_clean_no_dups = df_prods_clean.drop_duplicates()

df_prods_clean_no_dups.shape

df_prods_clean_no_dups.to_csv(os.path.join(path, '02 Data','Prepared Data', 'products_checked.csv'))

# 4.5 TASK

# 1. consistency checks for df_prods already run in the exercise above

# 2. Run df.describe() function on your df_prods dataframe
df_prods.describe()


Looking at the table we can see that all min values have
the same value 1, this is something we should look into. Aslo
when looking at the max values we can see an unusual price 9999.
We should also look into this value.

# 3. Check for mix-typed data in df_ords
for col in df_ords.columns.tolist():
  weird = (df_ords[[col]].applymap(type) != df_ords[[col]].iloc[0].apply(type)).any(axis = 1)
  if len (df_ords[weird]) > 0:
    print (col)

NO mixed-typed data found in df_ords dataframe

# 5. check missing values in df_ords
df_ords.isnull().sum()

206209 missing values found in df_ords dataframe from days_since_prior_order column 

# 6. Address missing values
df_nan1 = df_ords[df_ords['days_since_prior_order'].isnull() == True]

df_nan1

df_ords.shape

The total number of columns in this dataframe is 3421083, while the total number of missing values in the column days_since_prior_order is 206209. The fact that those values are missing indicates that there was only one order made and there there was no data related to when the prior order was made. So in this case I will choose to substitue those missing values with a zero.

# imputing zero values to missing value cells
df_ords['days_since_prior_order'].fillna(0, inplace=True)

df_ords

# checkin if function worked
df_ords.isnull().sum()

No missing values

# 7. Check for duplicate values in df_ords
df_duplic = df_ords[df_ords.duplicated()]

df_duplic

No duplicates found in df_ords dataframe

# 9. Export final df_prods (already exported at the end of the exercise) and df_ords as "csv" files
df_ords_clean_no_dups.to_csv(os.path.join(path, '02 Data','Prepared Data', 'orders_checked.csv'))
