==> 6. Instacart Combining and exporting data (part 1).ipynb <==
# Import libraries

import pandas as pd
import numpy as np
import os

# Import data
df_ords= pd.read_csv(r'/Users/jlsanabria77/Desktop/17-02.2021 Instacat Basket Analysis 3/02 Data/Original Data/orders.csv', index_col = False)

# Import data
df_ords_prior = pd.read_csv(r'/Users/jlsanabria77/Desktop/17-02.2021 Instacat Basket Analysis 3/02 Data/Prepared Data/order_products__prior.csv', index_col = False)

# Turning folderpath into a string
path = r'/Users/jlsanabria77/Desktop/17-02.2021 Instacat Basket Analysis 3'

# Check the output
df_ords_prior.head()

df_ords_prior.shape


df_ords.shape

# Merge both dataframes
df_merged_large = df_ords.merge(df_ords_prior, on = 'order_id', indicator = True)

df_merged_large.head()

df_merged_large['_merge'].value_counts()

df_merged_large['_merge'].value_counts()

df_merged_large.drop(['_merge'], axis=1, inplace = True)

df_merged_large.head()

# Export data to pkl

df_merged_large.to_pickle(os.path.join(path, '02 Data','Prepared Data', 'orders_products_combined.pkl'))
