==> 6. Instacart Combining and exporting data (part2).ipynb <==
# Import libraries

import pandas as pd
import numpy as np
import os

# 3. Import data
df_ords_prods_combined = pd.read_pickle(r'/Users/jlsanabria77/Desktop/17-02.2021 Instacat Basket Analysis 3/02 Data/Prepared Data/orders_products_combined.pkl')

# Turning folderpath into a string
path = r'/Users/jlsanabria77/Desktop/17-02.2021 Instacat Basket Analysis 3'

# 4. Check the shape of the imported data
df_ords_prods_combined.shape

# 4.1 Check it is the same as the exported data in the exercise

-From the exercise notebook. It has the same size

df_merged_large['_merge'].value_counts()
Out[16]:
both          32434489
right_only           0
left_only            0
Name: _merge, dtype: int64

# 5. combine orders_products_combined with products dataset

# 5.2 Import products dataset
df_prods_checked = pd.read_csv(r'/Users/jlsanabria77/Desktop/17-02.2021 Instacat Basket Analysis 3/02 Data/Prepared Data/products_checked.csv', index_col = False)

# I will combine both datasets using a merge. Before I will check their shape to see how columns look

df_prods_checked.head()

df_ords_prods_combined.head()

# 5.3 By looking at both df I can see only product_id column can be used as a link to create a merge.
df_orders_products_merged = df_ords_prods_combined.merge(df_prods_checked, on = 'product_id', indicator = True)

df_orders_products_merged.head()

# using the "merge flag" we can see that there is a full match of values
df_orders_products_merged['_merge'].value_counts()

#  6. For the size of this merged dataframe I will choose pickle to export

df_orders_products_merged.to_pickle(os.path.join(path, '02 Data','Prepared Data', 'orders_products_merged.pkl'))
