==> 7. Instacart Deriving New Variables.ipynb <==
# Import libraries
import pandas as pd
import numpy as np
import os
import warnings 
warnings.filterwarnings('ignore')

ords_prods_merge = pd.read_pickle(r'/Users/jlsanabria77/Desktop/17-02.2021 Instacat Basket Analysis 3/03 Scripts/02 Data/Prepared Data/orders_products_merged1.pkl')

df = ords_prods_merge[:1000000]

df.shape

df.head()

# Define the Function

def price_label(row):

  if row['prices'] <= 5:
    return 'Low-range product'
  elif (row['prices'] > 5) and (row['prices'] <= 15):
    return 'Mid-range product'
  elif row['prices'] > 15:
    return 'High range'
  else: return 'Not enough data'

# Apply the Function
df['price_range'] = df.apply(price_label, axis=1)

df['price_range'].value_counts(dropna = False)

df['prices'].max()

df.loc[df['prices'] > 15, 'price_range_loc'] = 'High-range product'

df.loc[(df['prices'] <= 15) & (df['prices'] > 5), 'price_range_loc'] = 'Mid-range product'

df.loc[df['prices'] <= 5, 'price_range_loc'] = 'Low-range product'

df['price_range_loc'].value_counts(dropna = False)

ords_prods_merge.loc[ords_prods_merge['prices'] > 15, 'price_range_loc'] = 'High-range product'

ords_prods_merge.loc[(ords_prods_merge['prices'] <= 15) & (df['prices'] > 5), 'price_range_loc'] = 'Mid-range product'

ords_prods_merge.loc[ords_prods_merge['prices'] <= 5, 'price_range_loc'] = 'Low-range product'

ords_prods_merge['price_range_loc'].value_counts(dropna = False)

for x in range(30, 45):
    print("My age is %d" % (x))

ords_prods_merge['order_dow'].value_counts(dropna = False)

result = []

for value in ords_prods_merge["order_dow"]:
  if value == 0:
    result.append("Busiest day")
  elif value == 4:
    result.append("Least busy")
  else:
    result.append("Regularly busy")

result

ords_prods_merge['busiest_day'] = result

ords_prods_merge['busiest_day'].value_counts(dropna = False)

# TASK 4.7

2. Change the Column to "Busiest days"

# we will find the two busiest and slowest days.
ords_prods_merge['order_dow'].value_counts(dropna = False)

-The two busiest days are 0 (Sunday) & 1 (Monday)

-The two slowest days are 3 (Wednesday) & 4 (Thursday)

# Create a new column with 'two_busiest' and 'two_slowest'
result = []

for value in ords_prods_merge["order_dow"]:
  if value == 0 or value == 1:
    result.append("Busiest days")
  elif value == 3 or value == 4:
    result.append("Slowest days")
  else:
    result.append("Regularly busy")

result

ords_prods_merge['busiest_days'] = result

ords_prods_merge['busiest_days'].value_counts(dropna = False)

# Check the new column has been created.
ords_prods_merge.head()

### 3. By comparing both frequencies from both columns we can see that the sum of Sunday and Monday equals the frequency for "busiest days"

### 4. Create a new column called "busiest_period_of_day"

# First we will print the frequency to find out which are the busiest hours of the day.
ords_prods_merge['order_hour_of_day'].value_counts(dropna = False)

By looking at the frequency we can see that the period with most orders is is from 9h to 16h
The period with fewest orders is between midnight to 6am.

# We will create a new column containing these labels.
result = []

for value in ords_prods_merge["order_hour_of_day"]:
  if value >= 9  and value <= 16:
    result.append("Most Orders")
  elif value >= 0 and value <= 6:
    result.append("Fewest Orders")
  else:
    result.append("Average Orders")

# Let's check the values
result

# we create a new column "busiest_period_of_day"
ords_prods_merge['busiest_period_of_day'] = result

### 5. Print the frequency for this new column.

# We print the frequency for this column to check the results
ords_prods_merge['busiest_period_of_day'].value_counts(dropna = False)

# Export dataframe as pikle file to "Prepared Data" folder
ords_prods_merge.to_pickle(os.path.join(r'/Users/jlsanabria77/Desktop/17-02.2021 Instacat Basket Analysis 3/02 Data/Prepared Data/ords_prods_merge.pkl'))
