==> 8. Instacart Grouping and Aggregating (part1).ipynb <==
# Import libraries

import pandas as pd
import numpy as np
import os

ords_prods_merge = pd.read_pickle(r'/Users/jlsanabria77/Desktop/17-02.2021 Instacat Basket Analysis 3/02 Data/Prepared Data/ords_prods_merge.pkl')

ords_prods_merge.groupby('department_id').agg({'order_number': ['mean']})

3. When analyzing the results there are few aspects worth to mention. First the enire dataframe produces 21 departments while the subset only showed eight. Also When looking at the mean from each deapartment we can see that those are slightly lower. For example, department_id 7 has a mean of 17.472355 in the subset but a mean of 17.225802 in the entire dataframe.

2. Create a Loyalty flag

# 1st. I will split the data into groups based on the “user_id” column.
# 2nd. I will apply the transform() function on the “order_number” column to generate the maximum orders for each user.
# 3rd. I will create a new column, “max_order,” into which I will place the results of the aggregation.

ords_prods_merge['max_order'] = ords_prods_merge.groupby(['user_id'])['order_number'].transform(np.max)

# Check the new column has been created
ords_prods_merge.head()

Create a flag that assigns a “loyalty” label to a user ID based on its corresponding max order value

ords_prods_merge.loc[ords_prods_merge['max_order'] > 40, 'loyalty_flag'] = 'Loyal customer'

ords_prods_merge.loc[(ords_prods_merge['max_order'] <= 40) & (ords_prods_merge['max_order'] > 10), 'loyalty_flag'] = 'Regular customer'

ords_prods_merge.loc[ords_prods_merge['max_order'] <= 10, 'loyalty_flag'] = 'New customer'

ords_prods_merge['loyalty_flag'].value_counts(dropna = False)

5. Determine whether prices of products purchased by loyal customers differ from those purchased by regular or new customers

ords_prods_merge.head()

# By using the groupby I will do multiple aggregations to find out the maximum, minimum and average price of products bought by these three type of customers
ords_prods_merge.groupby('loyalty_flag').agg({'prices': ['mean','max','min']})

We can see that all three groups buy from the cheapest (1.0) to the most expensive (99999.0) product. 
Also we observe that Loyal Customers buy on average less expensive products than the other two groups. 

6. Create a spending flag for each user based on the average price across all their orders using the following criteria:

# 1st. I will create a column of mean spend by user_id
ords_prods_merge['mean_spend'] = ords_prods_merge.groupby(['user_id'])['prices'].transform(np.mean)

# 2nd. I will create a spend flag for "Low Spender"
ords_prods_merge.loc[ords_prods_merge['mean_spend'] < 10, 'spend_flag'] = 'Low Spender'

# 3rd. I will create a spend flag for "High Spender"
ords_prods_merge.loc[ords_prods_merge['mean_spend'] >= 10, 'spend_flag'] = 'High Spender'

# 4th. I will check all values 
ords_prods_merge['spend_flag'].value_counts(dropna = False)

7. Determine frequent versus non-frequent customers. Create an order frequency flag that marks the regularity of a user’s ordering behavior according to the median in the “days_since_prior_order” column.

# 1st. I will create a column of median for days_since_prior_order column
ords_prods_merge['median_ordering'] = ords_prods_merge.groupby(['user_id'])['days_since_prior_order'].transform(np.median)

# 2nd. I will create a frequency flag for "Non-frequent-customer"
ords_prods_merge.loc[ords_prods_merge['median_ordering'] > 20, 'frequency_flag'] = 'Non-frequent customer'

# 3th. I will create a frequency flag for "Regular customer"
ords_prods_merge.loc[(ords_prods_merge['median_ordering'] > 10) & (ords_prods_merge['median_ordering'] <= 20), 'frequency_flag'] = 'Regular customer'

# 4th. I will create a frequenct flag for "Frequent customer"
ords_prods_merge.loc[ords_prods_merge['median_ordering'] <= 10, 'frequency_flag'] = 'Frequent customer'

# 5th. I will check all values
ords_prods_merge['frequency_flag'].value_counts(dropna = False)

# Export dataframe as pikle file to "Prepared Data" folder
ords_prods_merge.to_pickle(os.path.join(r'/Users/jlsanabria77/Desktop/17-02.2021 Instacat Basket Analysis 3/02 Data/Prepared Data/ords_prods_merge.pkl'))
