==> 9. Instacart visualization (part 1).ipynb <==
# Import analysis libraries
import pandas as pd
import numpy as np
import os
import matplotlib.pyplot as plt
import seaborn as sns
import scipy

# Import customers dataframe
df = pd.read_csv(r'/Users/jlsanabria77/Desktop/17-02.2021 Instacat Basket Analysis 3/02 Data/Original Data/customers.csv',index_col = False)

# create a path to make following steps easier
path = r'/Users/jlsanabria77/Desktop/17-02.2021 Instacat Basket Analysis 3'

# 4. Wrangle data.

# check column names
df.head()

I will drop 'First Name',Surnam' and 'Gender' as they do not add anything to the analysis. After checking the rest of the task (part2) requirements and the task for exercise 10 it looks like all the remaining columns will be needed for the final analyis therefore only those three can be dropped.

I also will change 'n_dependants' for 'number_of_dependants'

'fam_status' for 'family_status' and 'income' for 'annual_income'

# Dropping columns
df = df.drop(columns = ['First Name','Surnam','Gender'])

# Changing column names
df.rename(columns = {'n_dependants':'number_of_dependants','fam_status':'family_status','income':'annual_income'}, inplace = True)

# Check dropped and new named columns
df.head()

# 5. Check for Quality and Consistency

# Check and convert any mixed-type data

for col in df.columns.tolist():
  weird = (df[[col]].applymap(type) != df[[col]].iloc[0].apply(type)).any(axis = 1)
  if len (df[weird]) > 0:
    print (col)

By using the the above comand we see there isn't mixed-type data in this dataframe

# Finding and addressing any missing values

df.isnull().sum()

By using the above comand we see there aren't any missing values

# Find and addressing duplicates
# We create a subset containig only duplicates

df_dups = df[df.duplicated()]

df_dups

By using the above comand we see there aren't any duplicates (the subset has no duplicate rows)

# 6. Combine customer data with rest of prepared data

# First we will export prepared customer dataframe

df.to_csv(os.path.join(path, '02 Data','Prepared Data','customers_prepared.csv'))

Import Already Prepared data and new prepared data (customers_prepared.csv)

# Import Instacart prepared data 
df1 = pd.read_pickle(r'/Users/jlsanabria77/Desktop/17-02.2021 Instacat Basket Analysis 3/02 Data/Prepared Data/ords_prods_merge8.pkl')

df1.head()

# By looking at both Df, only user_id is a link to create a merge
df3_merged = df1.merge(df, on = 'user_id', indicator = True)

# Export the new dataframe
df3_merged.to_pickle(os.path.join(r'/Users/jlsanabria77/Desktop/17-02.2021 Instacat Basket Analysis 3/02 Data/Prepared Data/ords_prods_merged88.pkl'))

df3_merged.head()

We can see all rows from the new merged dataframe

# using the "merge flag" we can see that there is a full match of values
df3_merged['_merge'].value_counts()
